pyproject.toml
[tool.poetry]
name = "test-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = ">=2.0.0,<3.0.0"
py-cord = "^2.0.0b4"
python-dotenv = "^0.19.2"
replit = "^3.2.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
